castle-api/pyproject.toml
[project]
name = "castle-api"
version = "0.1.0"
description = "Castle API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "castle-cli",
]

[tool.uv.sources]
castle-cli = { path = "../cli", editable = true }

[project.scripts]
castle-api = "castle_api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/castle_api"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[tool.ruff.lint.isort]
known-first-party = ["castle_api"]
